Move tests to other file
